+++ src/components/TodoContents.tsx
@@ -1,16 +1,17 @@
 import styled from "@emotion/styled";
 import TodoInput from "components/TodoInput";
 import TodoList from 'components/TodoList';
-import React, { useState } from 'react';
+import React, {useState} from 'react';
+import {TodoContext} from "../contexts/ToDoContext";
 
 const Container = styled.div`
-    display: block;
-    align-items: center;
-    justify-content: center;
-    width: 100%;
-    min-width: 500px;
-    padding: 0 8px;
-    margin-right: 10px;
+  display: block;
+  align-items: center;
+  justify-content: center;
+  width: 100%;
+  min-width: 500px;
+  padding: 0 8px;
+  margin-right: 10px;
 `;
 
 export interface TodoItem {
@@ -20,30 +21,13 @@ export interface TodoItem {
 
 const TodoContents = () => {
     const [todos, setTodos] = useState<TodoItem[]>([]);
-
-    const addTodo = (todo: string) => {
-        if (todo.trim() === '') {
-            return;
-        }
-
-        const newTodoItem: TodoItem = {
-            id: Date.now().toString(),
-            text: todo,
-        };
-
-        setTodos([...todos, newTodoItem]);
-        console.log("!!!!!", todos);
-    };
-
-    const removeTodo = (id: string) => {
-        setTodos(todos.filter(todo => todo.id != id));
-    };
-    
     return (
-        <Container>
-            <TodoInput addTodo = {addTodo} />
-            <TodoList todos = {todos} removeTodo = {removeTodo} />
-        </Container>
+        <TodoContext.Provider value={{todos, setTodos}}>
+            <Container>
+                <TodoInput/>
+                <TodoList/>
+            </Container>
+        </TodoContext.Provider>
     );
 }
