+++ src/components/TodoList.tsx
@@ -1,42 +1,44 @@
 import styled from "@emotion/styled";
 // import TodoItem from 'components/TodoContents'
 import Button from 'components/Button';
-import { TodoItem } from "./TodoContents";
+import {useGlobalContext} from "../contexts/ToDoContext";
 // import React, { useState } from 'react';
 
 const Container = styled.div`
-    align-items: center;
-    justify-content: center;
-    padding: 20px 0;
+  align-items: center;
+  justify-content: center;
+  padding: 20px 0;
 `;
 
 const List = styled.div`
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    padding: 3px 0;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  padding: 3px 0;
 `;
 
 const Todo = styled.span`
-    width: 500px;
-    min-width: 300px;
-    align-items: center;
-    margin-right: 10px;
-    padding: 10px 0;
+  width: 500px;
+  min-width: 300px;
+  align-items: center;
+  margin-right: 10px;
+  padding: 10px 0;
 `;
 
-interface Props {
-    readonly removeTodo?: (id: string) => void;
-    readonly todos?: TodoItem[];
-}
+const TodoList = () => {
 
-const TodoList = ({ todos, removeTodo }: Props) => {
+    const {todos, setTodos} = useGlobalContext()
+
+    const removeTodo = (id: string) => {
+        setTodos(todos.filter(todo => todo.id != id));
+    };
+
     return (
         <Container>
             {todos?.map((todo) => (
-                <List key = {todo.id}>
+                <List key={todo.id}>
                     <Todo>{todo.text}</Todo>
-                    <Button label = "Del" onClick = {() => (removeTodo && removeTodo(todo.id))}/>
+                    <Button label="Del" onClick={() => removeTodo(todo.id)}/>
                 </List>
             ))}
         </Container>
